fixed MSVC2022 issues
* enabled all warnings (fixed issues)
* work around co_yield visit issue
* upgraded compiler options

+++ src/parser.lib/parser/LineErrorReporter.h
@@ -112,7 +112,7 @@ inline auto escapeSourceLine(strings::View view, ViewMarkers viewMarkers) -> Esc
         auto str = escaped.str();
         output += strings::String{str.data(), str.data() + str.size()};
         begin = input.end();
-        offset += str.length();
+        offset += static_cast<int>(str.length());
     };
 
     for (auto d : strings::utf8Decode(view)) {
@@ -143,8 +143,8 @@ inline auto escapeSourceLine(strings::View view, ViewMarkers viewMarkers) -> Esc
         auto i = 0;
         for (const auto& vm : viewMarkers) {
             auto& m = markers[i];
-            m.start = vm.begin() - view.begin();
-            m.length = vm.byteCount().v;
+            m.start = static_cast<int>(vm.begin() - view.begin());
+            m.length = static_cast<int>(vm.byteCount().v);
             i++;
         }
         return EscapedMarkers{to_string(view), std::move(markers)};
@@ -165,9 +165,10 @@ void reportDecodeErrorMarkers(
     for (auto& m : escapedMarkers) highlights.emplace_back(Marker{m, {}});
 
     auto doc = Document{
-        {Paragraph{(viewMarkers.size() == 1) ? String{"The UTF8-decoder encountered an invalid encoding"}
-                                             : String{"The UTF8-decoder encountered multiple invalid encodings"},
-                   {}},
+        {Paragraph{
+             (viewMarkers.size() == 1) ? String{"The UTF8-decoder encountered an invalid encoding"}
+                                       : String{"The UTF8-decoder encountered multiple invalid encodings"},
+             {}},
          SourceCodeBlock{escapedLines, highlights, String{}, line}}};
 
     auto expl = Explanation{String("Invalid UTF8 Encoding"), doc};
@@ -268,8 +269,9 @@ void reportNewline(
         auto highlights = Highlights{};
         for (auto& m : escapedMarkers) highlights.emplace_back(Marker{m, {}});
 
-        auto doc = Document{{Paragraph{String{"The indentation mixes tabs and spaces."}, {}},
-                             SourceCodeBlock{escapedLines, highlights, String{}, text::Line{nli.position.line.v - 1}}}};
+        auto doc = Document{
+            {Paragraph{String{"The indentation mixes tabs and spaces."}, {}},
+             SourceCodeBlock{escapedLines, highlights, String{}, text::Line{nli.position.line.v - 1}}}};
 
         auto expl = Explanation{String("Mixed Indentation Characters"), doc};
 
@@ -314,8 +316,9 @@ void reportStringLiteral(
         using Kind = scanner::StringError::Kind;
         switch (err.kind) {
         case Kind::EndOfInput: {
-            auto doc = Document{{Paragraph{String{"The string was not terminated."}, {}},
-                                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
+            auto doc = Document{
+                {Paragraph{String{"The string was not terminated."}, {}},
+                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
             auto expl = Explanation{String("Unexpected end of input"), doc};
             auto d = Diagnostic{Code{String{"rebuild-lexer"}, 10}, Parts{expl}};
             context.reportDiagnostic(std::move(d));
@@ -326,32 +329,36 @@ void reportStringLiteral(
             break;
         }
         case Kind::InvalidEscape: {
-            auto doc = Document{{Paragraph{String{"These Escape sequences are unknown."}, {}},
-                                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
+            auto doc = Document{
+                {Paragraph{String{"These Escape sequences are unknown."}, {}},
+                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
             auto expl = Explanation{String("Unkown escape sequence"), doc};
             auto d = Diagnostic{Code{String{"rebuild-lexer"}, 11}, Parts{expl}};
             context.reportDiagnostic(std::move(d));
             break;
         }
         case Kind::InvalidControl: {
-            auto doc = Document{{Paragraph{String{"Use of invalid control characters. Use escape sequences."}, {}},
-                                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
+            auto doc = Document{
+                {Paragraph{String{"Use of invalid control characters. Use escape sequences."}, {}},
+                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
             auto expl = Explanation{String("Unkown control characters"), doc};
             auto d = Diagnostic{Code{String{"rebuild-lexer"}, 12}, Parts{expl}};
             context.reportDiagnostic(std::move(d));
             break;
         }
         case Kind::InvalidDecimalUnicode: {
-            auto doc = Document{{Paragraph{String{"Use of invalid decimal unicode values."}, {}},
-                                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
+            auto doc = Document{
+                {Paragraph{String{"Use of invalid decimal unicode values."}, {}},
+                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
             auto expl = Explanation{String("Invalid decimal unicode"), doc};
             auto d = Diagnostic{Code{String{"rebuild-lexer"}, 13}, Parts{expl}};
             context.reportDiagnostic(std::move(d));
             break;
         }
         case Kind::InvalidHexUnicode: {
-            auto doc = Document{{Paragraph{String{"Use of invalid hexadecimal unicode values."}, {}},
-                                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
+            auto doc = Document{
+                {Paragraph{String{"Use of invalid hexadecimal unicode values."}, {}},
+                 SourceCodeBlock{escapedLines, highlights, String{}, sl.position.line}}};
             auto expl = Explanation{String("Invalid hexadecimal unicode"), doc};
             auto d = Diagnostic{Code{String{"rebuild-lexer"}, 14}, Parts{expl}};
             context.reportDiagnostic(std::move(d));
@@ -392,22 +399,25 @@ void reportNumberLiteral(
                 reportDecodeErrorMarkers(nl.position.line, tokenLines, viewMarkers, context);
             },
             [&, &escapedLines = escapedLines](const scanner::NumberMissingExponent&) {
-                auto doc = Document{{Paragraph{String{"After the exponent sign an actual value is expected."}, {}},
-                                     SourceCodeBlock{escapedLines, highlights, String{}, nl.position.line}}};
+                auto doc = Document{
+                    {Paragraph{String{"After the exponent sign an actual value is expected."}, {}},
+                     SourceCodeBlock{escapedLines, highlights, String{}, nl.position.line}}};
                 auto expl = Explanation{String("Missing exponent value"), doc};
                 auto d = Diagnostic{Code{String{"rebuild-lexer"}, 20}, Parts{expl}};
                 context.reportDiagnostic(std::move(d));
             },
             [&, &escapedLines = escapedLines](const scanner::NumberMissingValue&) {
-                auto doc = Document{{Paragraph{String{"After the radix sign an actual value is expected."}, {}},
-                                     SourceCodeBlock{escapedLines, highlights, String{}, nl.position.line}}};
+                auto doc = Document{
+                    {Paragraph{String{"After the radix sign an actual value is expected."}, {}},
+                     SourceCodeBlock{escapedLines, highlights, String{}, nl.position.line}}};
                 auto expl = Explanation{String("Missing value"), doc};
                 auto d = Diagnostic{Code{String{"rebuild-lexer"}, 21}, Parts{expl}};
                 context.reportDiagnostic(std::move(d));
             },
             [&, &escapedLines = escapedLines](const scanner::NumberMissingBoundary&) {
-                auto doc = Document{{Paragraph{String{"The number literal ends with an unknown suffix."}, {}},
-                                     SourceCodeBlock{escapedLines, highlights, String{}, nl.position.line}}};
+                auto doc = Document{
+                    {Paragraph{String{"The number literal ends with an unknown suffix."}, {}},
+                     SourceCodeBlock{escapedLines, highlights, String{}, nl.position.line}}};
                 auto expl = Explanation{String("Missing boundary"), doc};
                 auto d = Diagnostic{Code{String{"rebuild-lexer"}, 22}, Parts{expl}};
                 context.reportDiagnostic(std::move(d));
@@ -444,22 +454,25 @@ void reportIdentifierLiteral(
                 reportDecodeErrorMarkers(ol.position.line, tokenLines, viewMarkers, context);
             },
             [&, &escapedLines = escapedLines](const scanner::OperatorWrongClose&) {
-                auto doc = Document{{Paragraph{String{"The closing sign does not match the opening sign."}, {}},
-                                     SourceCodeBlock{escapedLines, highlights, String{}, ol.position.line}}};
+                auto doc = Document{
+                    {Paragraph{String{"The closing sign does not match the opening sign."}, {}},
+                     SourceCodeBlock{escapedLines, highlights, String{}, ol.position.line}}};
                 auto expl = Explanation{String("Operator wrong close"), doc};
                 auto d = Diagnostic{Code{String{"rebuild-lexer"}, 30}, Parts{expl}};
                 context.reportDiagnostic(std::move(d));
             },
             [&, &escapedLines = escapedLines](const scanner::OperatorUnexpectedClose&) {
-                auto doc = Document{{Paragraph{String{"There was no opening sign before the closing sign."}, {}},
-                                     SourceCodeBlock{escapedLines, highlights, String{}, ol.position.line}}};
+                auto doc = Document{
+                    {Paragraph{String{"There was no opening sign before the closing sign."}, {}},
+                     SourceCodeBlock{escapedLines, highlights, String{}, ol.position.line}}};
                 auto expl = Explanation{String("Operator unexpected close"), doc};
                 auto d = Diagnostic{Code{String{"rebuild-lexer"}, 31}, Parts{expl}};
                 context.reportDiagnostic(std::move(d));
             },
             [&, &escapedLines = escapedLines](const scanner::OperatorNotClosed&) {
-                auto doc = Document{{Paragraph{String{"The operator ends before the closing sign was found."}, {}},
-                                     SourceCodeBlock{escapedLines, highlights, String{}, ol.position.line}}};
+                auto doc = Document{
+                    {Paragraph{String{"The operator ends before the closing sign was found."}, {}},
+                     SourceCodeBlock{escapedLines, highlights, String{}, ol.position.line}}};
                 auto expl = Explanation{String("Operator not closed"), doc};
                 auto d = Diagnostic{Code{String{"rebuild-lexer"}, 32}, Parts{expl}};
                 context.reportDiagnostic(std::move(d));
@@ -499,11 +512,12 @@ void reportUnexpectedCharacter(
     for (auto& m : escapedMarkers) highlights.emplace_back(Marker{m, {}});
 
     auto doc = Document{
-        {Paragraph{(viewMarkers.size() == 1)
-                       ? String{"The tokenizer encountered a character that is not part of any Rebuild language token."}
-                       : String{"The tokenizer encountered multiple characters that are not part of any Rebuild "
-                                "language token."},
-                   {}},
+        {Paragraph{
+             (viewMarkers.size() == 1)
+                 ? String{"The tokenizer encountered a character that is not part of any Rebuild language token."}
+                 : String{"The tokenizer encountered multiple characters that are not part of any Rebuild "
+                          "language token."},
+             {}},
          SourceCodeBlock{escapedLines, highlights, String{}, uc.position.line}}};
 
     auto expl = Explanation{String("Unexpected characters"), doc};
@@ -527,8 +541,9 @@ void reportUnexpectedColon(
     auto highlights = Highlights{};
     for (auto& m : escapedMarkers) highlights.emplace_back(Marker{m, {}});
 
-    auto doc = Document{{Paragraph{String{"The colon cannot be the only token on a line."}, {}},
-                         SourceCodeBlock{escapedLines, highlights, String{}, uc.position.line}}};
+    auto doc = Document{
+        {Paragraph{String{"The colon cannot be the only token on a line."}, {}},
+         SourceCodeBlock{escapedLines, highlights, String{}, uc.position.line}}};
 
     auto expl = Explanation{String("Unexpected colon"), doc};
 
@@ -593,8 +608,9 @@ void reportUnexpectedTokenAfterEnd(
     auto highlights = Highlights{};
     for (auto& m : escapedMarkers) highlights.emplace_back(Marker{m, {}});
 
-    auto doc = Document{{Paragraph{String{"After end no more tokens are allowed."}, {}},
-                         SourceCodeBlock{escapedLines, highlights, String{}, utae.position.line}}};
+    auto doc = Document{
+        {Paragraph{String{"After end no more tokens are allowed."}, {}},
+         SourceCodeBlock{escapedLines, highlights, String{}, utae.position.line}}};
 
     auto expl = Explanation{String("Unexpected tokens after end"), doc};
 
@@ -625,8 +641,9 @@ void reportUnexpectedBlockEnd(
     auto highlights = Highlights{};
     for (auto& m : escapedMarkers) highlights.emplace_back(Marker{m, {}});
 
-    auto doc = Document{{Paragraph{String{"The end keyword is only allowed to end blocks"}, {}},
-                         SourceCodeBlock{escapedLines, highlights, String{}, ube.position.line}}};
+    auto doc = Document{
+        {Paragraph{String{"The end keyword is only allowed to end blocks"}, {}},
+         SourceCodeBlock{escapedLines, highlights, String{}, ube.position.line}}};
 
     auto expl = Explanation{String("Unexpected block end"), doc};
 
@@ -648,8 +665,9 @@ void reportMissingBlockEnd(
     auto highlights = Highlights{};
     for (auto& m : escapedMarkers) highlights.emplace_back(Marker{m, {}});
 
-    auto doc = Document{{Paragraph{String{"The block ended without the end keyword"}, {}},
-                         SourceCodeBlock{escapedLines, highlights, String{}, ube.position.line}}};
+    auto doc = Document{
+        {Paragraph{String{"The block ended without the end keyword"}, {}},
+         SourceCodeBlock{escapedLines, highlights, String{}, ube.position.line}}};
 
     auto expl = Explanation{String("Missing Block End"), doc};
